+++ src/components/swap/AssetSelector.tsx
@@ -13,8 +13,9 @@ import {
   isSui,
   Asset,
   ChainAsset,
+  ChainData,
 } from "@gardenfi/orderbook";
-import { assetInfoStore, ChainData } from "../../store/assetInfoStore";
+import { assetInfoStore } from "../../store/assetInfoStore";
 import { BTC, swapStore } from "../../store/swapStore";
 import { IOType, network } from "../../constants/constants";
 import { modalStore } from "../../store/modalStore";
@@ -340,7 +341,7 @@ export const AssetSelector: FC<props> = ({ onClose }) => {
                     <img
                       src={c.icon}
                       alt={c.name}
-                      className={`h-full max-h-5 w-full max-w-5 rounded-full`}
+                      className={`h-full max-h-5 w-full max-w-5`}
                     />
                     {hoveredChain === c.name && (
                       <ChainsTooltip