+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.1.9]
+
+### Fixed
+
+- Fixed chain and wallet logos.
+- Fixed Icon rounded issue in assets and chain, added new footer.
+- Fixed Native assets balances fetching.
+- Fixed X Link URL
+
 ## [1.1.8]
 
 ### Added
@@ -9,6 +18,7 @@
 - Added starknet support for Xverse wallet.
 - Implemented Swap bridge functionality.
 - Added Alpen Testnet support.
+- Add solver_id in Create Order Request.
 
 ### Fixed